+++ CHANGELOG.md
@@ -6,6 +6,8 @@ CHANGED:
 
 - Increased minimum PHP version requirement to 7.4 as property type declarations were added
 - Updated dev dependencies
+- Improved static analysis
+
 REMOVED:
 
 FIXED: